# Conflicts:
#	src/main/java/com/ajou/TwotoSix/unitTest2/repository/MockRepository.java
#	src/main/java/com/ajou/TwotoSix/unitTest2/service/MockService.java
#	src/test/java/com/ajou/TwotoSix/unitTest2/domain/StudentTest.java

+++ src/main/java/com/ajou/TwotoSix/unitTest2/domain/Student.java
@@ -14,5 +14,5 @@ public class Student {
     private String studentId;
     private int currentSemester;
     private String major;
-    private float GPA;
+    private double GPA;
 }
